moving blog components into their own folder
